allow nested directories
It's useful (for designers) to be able to group their icons into
separate categories. For example, in Sketch, symbols can be grouped
using the '/' character, and will appear in cascading submenus
within the app. But when these are exported, the SVG files will be
created in subdirectories.

This change compresses SVGs in nested subdirectories. A double-dash
(--) replaces the slash (/) in the IDs of the exported symbols.